Move macros to Model
