+++ src/main/java/jimmy/mcgymmy/model/Model.java
@@ -7,6 +7,7 @@ import javafx.collections.ObservableList;
 import jimmy.mcgymmy.commons.core.GuiSettings;
 import jimmy.mcgymmy.commons.core.index.Index;
 import jimmy.mcgymmy.model.food.Food;
+import jimmy.mcgymmy.model.macro.MacroList;
 
 /**
  * The API of the Model component.
