+++ src/main/java/jimmy/mcgymmy/model/macro/MacroList.java
@@ -1,0 +1,73 @@
+package jimmy.mcgymmy.model.macro;
+
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.HashSet;
+import java.util.List;
+import java.util.Map;
+import java.util.Set;
+
+import jimmy.mcgymmy.model.macro.exceptions.DuplicateMacroException;
+import jimmy.mcgymmy.logic.parser.PrimitiveCommandParser;
+
+/**
+ * Immutable container for macros. Ensures a macro's name is not taken before
+ * allowing it to be added.
+ */
+public class MacroList {
+    private final Map<String, Macro> macros;
+    private final Set<String> commandNames;
+
+    /**
+     * Creates a new MacroList
+     */
+    public MacroList() {
+        this.macros = new HashMap<>();
+        this.commandNames = new HashSet<>(PrimitiveCommandParser.getRegisteredCommands());
+        this.commandNames.add("macro");
+    }
+
+    /**
+     * Creates a new MacroList with pre-populated values.
+     * @param macros list of macros to populate this with.
+     * @throws DuplicateMacroException If there are any duplicate macros.
+     */
+    public MacroList(List<Macro> macros) throws DuplicateMacroException {
+        this.macros = new HashMap<>();
+        this.commandNames = new HashSet<>(PrimitiveCommandParser.getRegisteredCommands());
+        this.commandNames.add("macro");
+        for (Macro macro : macros) {
+            if (this.isMacroNameTaken(macro.getName())) {
+                throw new DuplicateMacroException();
+            }
+            this.macros.put(macro.getName(), macro);
+        }
+    }
+
+    public boolean hasMacro(String name) {
+        return this.macros.containsKey(name);
+    }
+
+    private boolean isMacroNameTaken(String name) {
+        return this.hasMacro(name) || commandNames.contains(name);
+    }
+
+    /**
+     * Returns a new MacroList with the new macro.
+     * @param newMacro the macro to add.
+     * @throws DuplicateMacroException if the macro's name has already been taken.
+     */
+    public MacroList withNewMacro(Macro newMacro) throws DuplicateMacroException {
+        List<Macro> newList = this.getAsList();
+        newList.add(newMacro);
+        return new MacroList(newList);
+    }
+
+    public Macro getMacro(String name) {
+        return this.macros.get(name);
+    }
+
+    public List<Macro> getAsList() {
+        return new ArrayList<>(this.macros.values());
+    }
+}
