+++ src/test/java/jimmy/mcgymmy/logic/macro/MacroTest.java
@@ -15,6 +15,7 @@ import jimmy.mcgymmy.model.Model;
 import jimmy.mcgymmy.model.ModelManager;
 import jimmy.mcgymmy.model.UserPrefs;
 import jimmy.mcgymmy.model.food.Food;
+import jimmy.mcgymmy.model.macro.Macro;
 
 public class MacroTest {
     private final Macro testMacro;
